Add files via upload
This SQL file contains a data related to Guftugu Publications. This data was not in the required form to draw various results or conclusions out of it. Therefore we cleaned the data according to our needs using the queries and made a new table with the cleaned data. Later on we visualize this data on Tableau. It helped us in finding different trends. Like in KPK, people mostly prefer to read Islamic books. On the other hand in Karachi and Lahore, people prefer to read books related to Python programming language, product management and blockchain. The trend for buying these kind of books is rising and it suggests that for the next quarter, the number of books sold will increase.